JewelryStoreAPI/JewelryStore.API/Controllers/LoginController.cs
using JewelryStore.API.Filters;
using JewelryStore.Data.Interfaces;
using JewelryStore.Models.LoginModels;
using Microsoft.AspNetCore.Mvc;
using Microsoft.Extensions.Configuration;
using Microsoft.IdentityModel.Tokens;
using System;
using System.Globalization;
using System.IdentityModel.Tokens.Jwt;
using System.Security.Claims;
using System.Text;
using System.Threading.Tasks;

namespace JewelryStore.API.Controllers
{
    [Route("api/[controller]")]
    [ApiController]
    public class LoginController : ControllerBase
    {
        private readonly ILoginService _loginService;
        private readonly IConfiguration _config;
        public LoginController(ILoginService loginService, IConfiguration config)
        {
            _loginService = loginService;
            _config = config;
        }

        [HttpPost("authenticate")]
        [ProducesResponseType(typeof(LoginResponse), 200)]
        [ProducesResponseType(400)]
        [ProducesResponseType(401)]
        [ProducesResponseType(500)]
        [ServiceFilter(typeof(ModelValidationAttribute))]
        public async Task<IActionResult> Login([FromBody] LoginRequest loginRequest)
        {
            LoginResponse response = await _loginService.ValidateUser(loginRequest);
            if (response.Status)
            {
                // generate token
                var tokenHandler = new JwtSecurityTokenHandler();
                var key = Encoding.ASCII.GetBytes(_config.GetValue<string>("TokenGenerator:SecurityKey"));
                var tokenDescriptor = new SecurityTokenDescriptor
                {
                    Subject = new ClaimsIdentity(new Claim[]
                    {
                    new Claim(ClaimTypes.Email, response.Email),
                    new Claim(ClaimTypes.NameIdentifier, response.UserId.ToString(CultureInfo.InvariantCulture)),
                    new Claim(ClaimTypes.Role,response.Role)
                    }),
                    Issuer = _config["TokenGenerator:Issuer"],
                    Audience = _config["TokenGenerator:Audience"],
                    Expires = DateTime.UtcNow.AddHours(1),
                    SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(key), SecurityAlgorithms.HmacSha256Signature)
                };
                var token = tokenHandler.CreateToken(tokenDescriptor);
                response.Token = tokenHandler.WriteToken(token);
            }
            return StatusCode(response.HttpStatusCode, response);
        }
    }
}